ChessWeb.GameView (ChessDocumentation v0.2.0)

Functions

__resource__()

The resource name, as an atom, for this view

form.html(assigns)

game_over?(game)

index.html(assigns)

new.html(assigns)

opponent(game, user_id)

player(game, user_id)

player_colour(conn, game)

render(template, assigns \\ %{})

Renders the given template locally.

show.html(assigns)

state(conn, game)

template_not_found(template, assigns)

@spec template_not_found(binary(), map()) :: no_return()

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.

turn_class(conn, game)

won_lost(conn, game)

your_turn?(conn, game)